House Pricing Dataset
Ask a home buyer to describe their dream house, and they probably won’t begin with the height of the basement ceiling or the proximity to an east-west railroad. But this playground competition’s dataset proves that much more influences price negotiations that the number of bedrooms or a white-picket fence. With 79 explanatory variables describing challenges you to predict the final price of each home.